Refactor pim447.c to use uint16_t for delta_x and delta_y variables

+++ src/pim447.c
@@ -130,8 +130,8 @@ static void pim447_work_handler(struct k_work *work) {
     uint8_t delta_x_raw = (uint8_t)buf[1] - (uint8_t)buf[0]; // Right - Left
     uint8_t delta_y_raw = (uint8_t)buf[3] - (uint8_t)buf[2]; // Down - Up
 
-    uint8_t delta_x = convert_speed(delta_x_raw) * MOVE_FACTOR;
-    uint8_t delta_y = convert_speed(delta_y_raw) * MOVE_FACTOR;
+    uint16_t delta_x = convert_speed(delta_x_raw) * MOVE_FACTOR;
+    uint16_t delta_y = convert_speed(delta_y_raw) * MOVE_FACTOR;
 
 
     bool sw_pressed = (buf[4] & MSK_SWITCH_STATE) != 0;
